GP-204 -> Alterado module-web para ter suporte em _functions

+++ test/integration/_functions/abstract-module.sh
@@ -11,7 +11,9 @@ function removeModule
     moduleUrl=$(moduleUrl "$module")
     modulePath=$(modulepath "$moduleUrl")
 
-    sudo rm -R $modulePath
+    if [ -d "$modulePath" ***REMOVED***; then
+    	sudo rm -R $modulePath
+    fi
 }
 
 function resetModule
